Restructure analysis into full module
